requirements.txt
keras==2.2.4
tensorflow-gpu==1.13.1
scipy==1.2.1
pillow
